packages/workstation_vita/requirements.txt
streamlit==1.42.2
requests==2.32.5
numpy==2.2.6
pandas==3.0.1
faiss-cpu==1.12.0
python-dotenv==1.2.1
dingtalk-stream==0.24.3
